+++ index.html
@@ -10,7 +10,7 @@
 </head>
 
 <body>
-  <script src="js/main.js"> </script>
+  <script src="js/main.js" type = "module"> </script>
   <main>
 
     <!-- Фильтрация изображений от других пользователей -->
